addressing the PR comments
